misc: fix T1M config sometimes ignored
If for whatever reason the file with the original TombATI settings
cannot be deleted, the game would read that file and save T1M settings
over and over again, essentially making it impossible to change the
configuration across launches.

This is now fixed by only saving T1M settings if we managed to remove
the legacy file.