WIP: add support for disable_healing_between_levels with restart level
This code works for a current game session. There are major issues though:
1. Loading a different saved game carries over the stored level start health
from the previous play session.
2. Health is not stored on save, so if Lara dies and restart level is selected
without a level completetion in the current game session, the level restarts with
Lara at 0 health and she instantly dies.
3. Any players with a game save from an old Tomb1Main version will not have
stored health for the start of the level. There will need to be a check and
probably just default to full health.