s/init: clear GameMemory
Explicitly clear the memory to 0. GCC does this by default, other
compilers don't. MSVC sets it to 0xcdcdcdcd for debugging purposes.
Sadly the trigger data in Folly has an invalid item number which is
usually fine because it gets an all null, the floor is null and so
nothing is called.