objects/scion: fix reviving the Scion
After shooting the Scion in The Great Pyramid, if the player steps on
the tile that activates it again, the Scion will get reactivated as the
trigger is repeatable. This has a side effect of enabling Scion3Control
calls again and the internal explosion counter will tick beyond the
value of 90, causing even more explosions.

Rather than merely deactivating the Scion, we need to actually discard it.

Closes #252