Remove Fan to Strip kludge
So there was a function that was call strip that would memmove triangles and draw them 1 by 1, which then called a function that would loop through them and copy them into the vertex buffer. This has been removed and new Fan call for fan data has been added which will copy the vertices into the buffer without needing a memmove. Although we could actually just remove the Fan construction in the code and make it a strip anyway.