New Draw Screen Box
This uses raw vetex data to draw both "rects" at once with things snapped to vertices so they are no cracks, and it uses float maths so both "halves" are the same width. Probably will add more, it would make sense to draw the black transparent rectangle in this function as well, so it is perfectly aligned.

This also has a the other half of the Strip->Fan patch up.