GFX_GL: fix a copy-paste error.
MSVC complains the use of macro GFX_GL_CheckError(), saying that it needs a parameter. By looking its declaration into src\gfx\gl\utils.h, I have seen that it is declared with a "void" inside. Probably, in the past GFX_GL_CheckError() was a function and the "void" was forgotten when it had been converted into a macro. GCC doesn't suffer much, but CL thinks instead that "void" is a required parameter. Hopefully, removing that thing made all compilers happy.