Fix support for 64-bit
When running Tomb1Main compiled as a 64-bit application, I got an error.
This happens because ANIM_STRUCT has a frame_ptr item which is a pointer and so it varies its size if it's compiled for an addressing depth larger than 32-bit.
I fixed the issue by adding a frame_ofs item, used for indexing g_AnimFrames[], and by filling all the items of ANIM_STRUCT by reading them one by one, in the same manner it has been done into function Level_LoadRooms().

I tested Tomb1Main with the demo levels (see issue #667) and under Windows, compiled as i686 (32-bit) and x64 (64-bit) and the engine worked fine on both.