Fix varargs for Log_Message()
On Linux, the engine crashes when printing the log messages.
This happens because the current code re-uses the same va_list variable on two calls to vprintf() and vfprintf().
Actually, this is not allowed.
For using the same information on multiple formatting functions, it is needed to create a copy of the primary va_list to a second one, by using va_copy().
After rewriting properly the Log_Message() function, the segmentation fault is gone.
Tested on both Linux and Windows builds.