src/game/music.c: Music_Stop() is never called.
I run the engine with the demo files and some custom audio music tracks and I discovered a strange bug into `src/game/music.c`.
Here:
https://github.com/rr-/Tomb1Main/blob/fcdf33adf0f634c10fdf3e768f7e11c0a5c1b9b9/src/game/music.c#L73
there is this piece of code:

```
    if (track == 0) {
        Music_Stop();
        return false;
    }
```

but it seems to be never called because previously at this line:
https://github.com/rr-/Tomb1Main/blob/fcdf33adf0f634c10fdf3e768f7e11c0a5c1b9b9/src/game/music.c#L65
there is this code:

```
    if (track <= 1) {
        return false;
    }
```

which returns to the caller without doing anything.
In my opinion, the block of code at line 65 must be deleted and line 73 should be changed as into this PR.

Fixed also some constants with proper MX_ values as requested.