s/input: remove redundant S_Input_KbdKey()
It is not clear to me the link between S_Input_KbdKey() and S_Input_Key().
They are both static and S_Input_Key() just calls S_Input_KbdKey() with the same parameter. This looks a redundant stuff. In this patch, I deleted S_Input_Key() and I renamed S_Input_KbdKey(), which requires the minimum number of changes.