Fixes screenshots
This makes the Framebuffer the context and then grabs the whole frame from it. At the moment this makes all screen shots 800x600 but does give us the entire frame. Before it was trying to grab the visible frame from the screens viewpoint however that is not where the image is located on the framebuffer so it was cutting the image off and grabbing a lot of black.