inject: fix music trigger setups in Folly (#946)
In St. Francis' Folly, the music trigger for track 3 in room 4 has been
moved behind the Neptune door. A new trigger for track 15 has also been
added once the 4 keys are used at the end of the level.
A new injection function to support moving rooms has been added. This
is used in Folly to move room 54 out of earshot so the player does not
hear the boulder that opens the door.

Resolves #865.