inject: fix Atlantean Stronghold room portals
Adds to the Atlantean Stronghold FD injection fixes by reconfiguring
the portals between rooms 74 and 12 such that the vertices are within
the room bounds in all cases. This amends the injection function for
this to allow targeting specific doors as there are 3 in this case.
Its other use case is in Palace Midas where there is only 1 portal.

Resolves #227.