build: always bump git version; remove init.c
Rebuilding the exe was not updating the git version shown inside the
game, which caused me some issues when I was working on the project
across two machines. This new approach forces meson to always reconsider
the current version, and rebuild relevant autogenerated files:
version.rc and init.c prior to linking the .exe.

Furthermore, the versioned init.c was removed. The existing users who
choose to build without Docker will continue to see fallback
information, but the mechanism was changed from a versioned file to a
special environment variable that's set by the Docker builds.

This change should require no action other than a clean project rebuild
on the part of the users.