ai: extend item drop support to all enemies
Added support for all creatures to carry and drop items; specifically,
exploding and water enemies on top of standard creatures. Items that
spawn in mid-air or in the water will fall to the floor.
Added a new enum for drop status for clarity, and updated the savegame
to ensure items are accurately re-positioned on load.

Resolves #1029.