config: remove obsolete macro postprocessor
This code used to matter for READ_/WRITE_ macros that needed to replace
strings in the form of "ui.text_scale" to "text_scale" for JSON. Now we
have full control over the JSON keys, so we no longer need that.

It would be best if we put the options in relevant JSON subsections, but
at the same time we need to be backwards compatible. I don't feel up to
the task just yet.