game: make game status modular
This removes the global game status and instead exposes functions for
setting, adding, removing and reverting status. Callers still need to
manage the call order, but it removes the various bitwise operations.
This resolves an issue where opening the inventory would not reverting
the status to GMS_IN_GAME to 0, which the check in Random_GetDraw looks
for. As a result, RNG was not progressing for such things as FX and
SFX.

Resolves #1070
Resolves #1074