game: make game status modular (#1075)
This removes the global game status and instead exposes functions for
getting and setting the status. This resolves an issue where opening
the inventory would not revert the status to GMS_IN_GAME to 0, which
the check in Random_GetDraw looks for. As a result, RNG was not
progressing for such things as FX and SFX.

Resolves #1070
Resolves #1074