missile: test range using components
This adjusts the missile damage range test to look at the individual
components of the difference between Lara and the FX in question,
rather than the Euclidean distance. There can be very large numbers
involved and this was leading to overflow; this overflow itself was
not detectable in release builds.

Resolves #1090.