movable_block: test animation index over state
LS_PP_READY refers to Lara's animation state when grabbing OR releasing
a pushblock. Prior to #1006, LF_PPREADY was an absolute frame index and
so the check for this would only pass after the grabbing animation.
With relative indexing, the check can pass in either animation, and so
holding action and forwards/backwards after releasing a block would
allow the test to pass and the block to move independently.

Replaced some remaining 1024 usage in this module with WALL_L const.

Resolves #1101.