lara: prevent look reset on target triggers
This ensures that if Lara remains on a trigger with a TO_TARGET entry
but no fixed camera, that she continues to look at the correct position
of said target. The camera's type and item is reset on each frame, but
the main camera update routine occurs after Lara's control routines,
hence we need to refer to the cached last item of the camera.

Resolves #1026.