camera: remove LOS check on fixed cameras
LOS checks are removed in TR3 onwards for fixed cameras, so implemented
here to fix the camera being thrown through doors that are yet to open.
This also improves other areas of the game, for example if Lara is
jumping down a slope with a low-ceiling and fixed camera, and if she
begins a level on a fixed camera but is just out of sight.

Resolves #954.