mac: update readme with first run before data copy
This allows Gatekeeper verification to pass and makes it drop
the quarantine filesystem xattrs, making it safe to copy over
the game data afterward.

If it's not done in this error, the OS will report the app
bundle as damaged instead.

Helps with issue #1359, however a more robust solution would
be to support loading game data from other locations.