tr2/box: use expanded box values
This adjusts TR2 boxes into the same structure as TR1, whereby values
are shifted once during level read so that the world coordinates can
be used directly in the box and camera modules. This includes a minor
update in TR1's camera shift clamp function for improved readability
matching TR2.