items: avoid interpolating specific items
This covers not interpolating items that are inactive, such as stopped
boulders. An item may still be active in the item list such that it
retains collision, but inanimate, sich as boulders or mines, so we need
to check both statuses. The TR1 test here has also been simplified to
match.

TR2 mines have a bad frame in the level data, so we avoid interpolating
these in all cases. The push button also has a single frame where a
mesh is rotated 180deg to simulate off/on, so this can't be
interpolated.