interpolation: detect large camera shifts
This updates camera interpolation in cases where there are large shifts
of any positional value, such as going to a fixed camera on the other
side of the world. In this case, we just commit the values rather than
interpolating each one, as this can result in the camera having settle
at the destination.