config/file: allow string integers
The config tool uses enums as the FPS property backing field in order to
limit user choice to either 30 or 60 FPS. The downside is that enum
types get serialized as strings, and TRX wouldn't recognize such values,
expected literal integers. The patch relaxes this requirement and makes
the config parser also accept numeric strings for integer-types.