trapdoor: fix startup problems
The cause was the bats - the code to fix bats embedded in room geometry
tries to find the nearest open spot to move the bats to. In the case of
the level "The Atlantean Armor" by Sabatu, it happened inside a room
with a trapdoor. At the time of invoking the embed fix, the trapdoor
didn't have its bounds ready and would crash the game with null pointer
access. It seems to be working fine otherwise.