tr2/flares: fix flicker in 60 fps
Resolves #2806. The OG mistakenly used the wrong arm (right instead of
left) to determine the flare light position in
`Lara_GetJointAbsPosition_I`. At 60 FPS, this mistake caused Lara's hand
position to appear significantly off from the expected location, which
coincidentally led to the flicker effect disappearing. While fixing the
bone did position Lara's hand correctly, it didn't bring back the
missing flicker at 60 FPS. In fact, it caused the flicker effect to
disappear at 30 FPS as well, making it static at both frame rates. It
turned out that although the game does randomize the position of the
flares' light source, it doesn't do so sufficiently. The primary cause
of the flickering was actually the randomness introduced by using the
incorrect bone. To restore the flicker effect to how it originally
appeared, I had to enhance the randomness of the flare's light position.