tr2/room: remove cinematic room initialiser
Room_InitCinematic was used to mark all cutscene level rooms as ready
to be drawn, and to also set the outside flag on each room. Since #2761
room drawing can calculate the visible rooms normally and there is no
need to rely on the preset. Testing has shown also that the outside
flag is not required in the OG cutscenes.