tr1/lara/cheat: avoid disabling mummy AI
The Qualopec mummy uses its data field to store a single short for its
head rotation. When Lara cheats to kill the mummy, LOT is called to
disable the slot, but this can lead to memory corruption as it expects
a regular CREATURE pointer. For drop items, this was resulting in the
carried item's spawn number to be set to -1. The carrier module would
try to re-spawn the item, but the rest of the data was in a corrupt
state, and therefore either a crash or strange outcomes would result.

Resolves #2957.