box: remove zone skipping on load
This removes TR2's zone skipping when reading level data based on the
loaded enemy type. LOT setup assigns the step value, and this is used
in the `BOX_ZONE` macro to retrieve the relevant zone index. Therefore
as the checks are linked to the only enemy types that use these zones,
the skipping was most likely in place for performance reasons.