sound: use similar sample data loading API
This makes TR1 and TR2 use a similar sample data loading API via the
sound module, and removes the dependency in the level modules on the
audio module directly. There remains some distinction with TR1 as it
also needs to reset ambient sounds separately after the samples are
loaded. Further merging will follow.