strings: use ring buffers
We introduced the static buffer approach to simplify the creation of
short-lived strings that need to be used right after they're generated.
However, there are cases where multiple strings must be created
simultaneously, such as a dialog title and its accompanying text
content. To accommodate this need, we convert the buffer into a ring
structure that allows having up to 8 valid strings at any single time.