camera: ensure environment in the control phase
Resolves #3305.

The underwater hum would play for at most 2 frames before getting reset
on certain devices. The reason for this is because of how
Sound_EndScene() works.

The idea is that looped sound sources need to Sound_Effect() their
samples on every logical frame to achieve continuous playback, otherwise
the samples will stop. Which is actually quite convenient. To achieve
this behavior, Sound_EndScene() checks if the looped samples are silent.
If they aren't, it syncs the parameters to the relevant output stream
and updates the samples volume to be silent – but keeps the other
parameters. In the following frame, the repeated call to Sound_Effect()
will either reset the volume back to some non-silent value, or if there
are no such calls, the second call to Sound_EndScene() will deem the
sample as no longer in use, stop the playback, and fully close the slot.

At the same time the call to Sound_Effect(SFX_UNDERWATER) would happen
inside M_EnsureEnvironment() which is part of the draw phase. Under
normal circumstances this is fine, even if we have 60 FPS. But if we
experience frame drops, this means the calls to Sound_Effect() and the
calls to Sound_EndScene() can go out of order , and the samples will
randomly restart. This is especially visible if the FPS is set to 60 but
the display is only capable of rendering 30 FPS.

The fix ended up being simple and was to just move the camera
environment sync to the control phase.