phase: more eager overriding of commands
No longer waits till a control() callback of a phase before issuing an
override command.

This helps with a scenario where the player issues a command like
`/play` during the opening FMV before the inventory ring was shown – now
the phase won't even try to start and as such will omit the music
trigger.