creature/common: retain only monks as allies
This removes Lara, Winston and the dying monk in Diving Area from the
list of ally types so that only allies who will fight other enemies
are retained. Those specific enemies now have their own array as
well to add an additional abstraction layer, and Winston is defined
as a special type in order to continue allowing him to be killed by
a direct console command only.

Resolves #3492.