lara/state/land: restore idle pose animation
This restores Lara's idle pose from early beta levels. It will trigger
after a period of inactivity (the number of seconds specified in
the config option) provided Lara is weapon and flare free, on land,
not in a vehicle, and not being nudged by an enemy. Regular movement
input will take her out of the state.

Resolves #3727.