collision: track room number when filling sides
In the original collision gathering routine, the room number could
change on each call to Room_GetSector, and hence affect the next side
being tested. Having moved this to a separate function, the base room
number was always used. This in turn fixed sticky corners by default,
and so the wall glitch mode option is now used to determine which
room to use.

Resolves #3957.