objects/movable_block: use default floor/ceiling when falling
This alters the movable block floor and ceiling functions to return the
default provided heights if the block is falling, which is effectively
equivalent to OG in this state. This avoids Lara being col shifted
on top of the block too early during her own collision routines,
and hence the killer check not being able to run in time.

Resolves #3981.