lara/common: disable interpolation in flip animations
This allows interpolation to be disabled in specific frames of some
of Lara's twist animations where she is drawn completely flat. The
neutral twist animation has been updated to perform the command one
frame later, which avoids Lara flipping twice with this change.

This is a temporary workaround until a better solution is found.

Resolves #4012.