items/actions: fix cutscene mesh swaps
This corrects mesh swap behaviour in cutscenes for Lara so that status
is more accurately tracked. The flip effects are used in the data as
toggles, but the difference when drawing/undrawing in game is that
Lara's gun status is used to track mesh status effectively. The mesh
pointers are now compared in cutscenes in each toggle call.

Resolves #4136.  Resolves #4142.  Resolves #4145.