lara/col/jump: fix wall glitch mode preventing normal animation
Wall glitches happen in TR1 and 2 when Lara is in her forward jump
animation and she has (effectively) reached the floor at exactly
the same time as hitting a wall. The Lara_Animate call is the key to
her being pushed into the wall. The previous fix for wall glitches
blocked this animation call on all occasions, and had the side effect
of Lara losing all forward momentum, the case here being when she hit
a springboard. The check now maintains momentum when Lara hasn't hit
a wall in front of her.

Resolves #4187.