lara/control: fix col update in extra anims
This avoids Lara's normal collision routines running if object
collision has switched her into the extra anim state. This doesn't
affect anywhere in OG because the routines with the same state IDs as
normal animations are harmless, but as we expand the extra animations
there is the risk of running mismatched routines that will further
alter Lara and break the extra animations.