rooms/geometry: fix sector lookup
In situations where a sector's floor is reasonably close to its
ceiling, and a triangular portal exists, the Y pos being tested in
Room_GetSector would likely traverse into the sector above because
it would compare with the untilted height on the sector. This adopts
the TR3 approach of ensuring that the precise ceiling and floor
values are tested for X and Z to guarantee that the correct sector is
returned. This in turn avoids Lara's collision using wrong sectors
as her height will more often than not push into the room above,
and hence cause problems with invisible walls and indeed the camera
misbehaving as the target moves between rooms rapidly.