output: separate camera and world view
For the longest time we were building our matrix stack starting off
`g_W2VMatrix`, which means that every matrix has had the view transform
already baked-in.

There are several issues with this.

- It's harder to reason about things.
- Techniques like billboarding shine best given clean world transforms.
  Separating it enables refactors that'll let us drop the warmup stage
  altogether – AND eliminate startup lag (at least on my machine).
- Moving lighting to the GPU is completely blocked without this.
- Also, surprise: `g_W2VMatrix` isn’t even a real view matrix! More on that later.

To split these matrices in the shader, I replicated `matrix.c`’s logic
on a new stack built from identity, tracked by `g_WMatrixPtr` (W for
World). This was done in 04fb407b87b0.

Next up I had to fix the view matrix, then adjust the mesh batcher to
take the world matrix alongside the camera*model premultiplied matrix,
and that was pretty much it. The rest was just making sure the UI
doesn't try to flirt with the world camera, as it's not built for a 3D
world. To that end, I chose to simplify things down to switching between
projection modes at runtime.

Also, I changed the render flush to only update `uMatPerspProjection`
and `uMatView`, and the rest of the Globals UBO to be updated only once
at the render start. These two matrices are the only parts that need
per-flush updates, since they can change when the player opens the
inventory ring.

(If this all sounds like a perfect excuse to drop `g_MatrixPtr` and only
go with `g_WMatrixPtr` everywhere – nope, it has to stay. It's still
essential for software clipping, which operates in the view coordinate
system from `g_MatrixPtr`, not the world space from `g_WMatrixPtr`.)

---

I mentioned `g_WMatrixPtr` was busted. It took me half a day to catch
it, but turns out the translation components in `g_W2VMatrix` were
storing the *raw camera position*.

A view matrix's whole job is to counter the camera's position and
rotation – essentially pulling the world back toward the origin so your
camera becomes the center of the scene. Meaning those translation
elements should aim toward 0,0,0, not *contain* the camera coordinates
themselves.

`Matrix_TranslateAbs` was *technically* doing the right operation, just
with all the wrong inputs. It took those bogus translation values,
treated them like valid view data, and re‑centered everything around the
camera.

The skybox followed suit, nuking its translation with:
```
g_MatrixPtr->_03 = 0;
g_MatrixPtr->_13 = 0;
g_MatrixPtr->_23 = 0;
```
and it took me a while why that was.

I fixed the view matrix to store the correct values and made it
available as `g_ViewPos`. In later commits I'll try to get rid of
`g_W2VMatrix`.