matrix: kill off g_W2VMatrix; introduce g_ViewPos
As explained in the previous commit, the g_W2VMatrix is confusing.
This commit removes its usages and for the code paths that need the
view pos, it offers it as a separate global.

Technically, g_ViewPos is 99% of the time the same as
g_Camera.interp.result.pos, but I wasn't ready to go after the remaining
1%.