rooms/draw: update water status for each item draw
This sets the output water status before drawing each item, to ensure
that the room that has been determined in control updates is the
one used for tinting. Sorting by room status has been reverted as
there were a few static mesh issues, such as the poles in Venice and
the life raft in The Deck. Injections are provided to further fix
z-fighting in Venice.