objects/switch: amend switch states
Level data defines 0 as switch state on and 1 as off. When testing
triggers, the check was originally done against the switch having
state LS_RUN to mean off, but this was fixed to use the correct switch
state name, but in turn inherited the wrong value.  The utility switch
state functions have been renamed slightly to reflect what they do
more specifically.

Resolves  #4456.